+++ crates/sncast/tests/e2e/declare.rs
@@ -16,7 +16,6 @@ use sncast::helpers::fee::FeeArgs;
 use starknet::core::types::TransactionReceipt::Declare;
 use starknet_types_core::felt::{Felt, NonZeroFelt};
 use std::fs;
-use tempfile::tempdir;
 use test_case::test_case;
 
 #[tokio::test]
@@ -614,43 +613,6 @@ async fn test_no_scarb_profile() {
             To see declaration details, visit:
             class: [..]
             transaction: [..]
-        "},
-    );
-}
-
-// TODO(#3118: Remove this test, once integration with braavos is restored
-#[tokio::test]
-async fn test_braavos_disabled() {
-    let contract_path = duplicate_contract_directory_with_salt(
-        CONTRACTS_DIR.to_string() + "/map",
-        "put",
-        "human_readable",
-    );
-    let tempdir = tempdir().expect("Failed to create a temporary directory");
-    let accounts_json_path = get_accounts_path("tests/data/accounts/accounts.json");
-    join_tempdirs(&contract_path, &tempdir);
-
-    let args = vec![
-        "--accounts-file",
-        &accounts_json_path,
-        "--account",
-        "braavos_incompatible",
-        "declare",
-        "--url",
-        URL,
-        "--contract-name",
-        "Map",
-    ];
-    let args = apply_test_resource_bounds_flags(args);
-
-    let snapbox = runner(&args).current_dir(tempdir.path());
-    let output = snapbox.assert().failure();
-
-    assert_stderr_contains(
-        output,
-        indoc! {r"
-        Error: Using incompatible Braavos accounts is disabled because they don't work with starknet 0.13.5.
-            Visit this link to read more: https://community.starknet.io/t/starknet-devtools-for-0-13-5/115495#p-2359168-braavos-compatibility-issues-3
         "},
     );
 }