+++ CHANGELOG.md
@@ -20,9 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Cast
 
-#### Updated
+#### Changed
 
 - OpenZeppelin account updated to v1.0.0 [preset](https://docs.openzeppelin.com/contracts-cairo/1.0.0/api/account#AccountUpgradeable)
+- `account create` command with `--type braavos` is temporarily disabled. Please use Braavos web wallet to create an account and later import it with `sncast account import`.
 
 ## [0.43.0] - 2025-05-09